Update Tricium analyzer for metrics to be compatible with recipes
This CL includes shifting code around and path-mangling to make the Tricium analyzer for metrics compatible with both future recipes and current tests.

Bug: 1004564

+++ go/src/infra/tricium/functions/metrics/histograms_check.go
@@ -143,55 +143,55 @@ const (
 func main() {
 	inputDir := flag.String("input", "", "Path to root of Tricium input")
 	outputDir := flag.String("output", "", "Path to root of Tricium output")
+	prevDir := flag.String("previous", "", "Path to directory with previous versions of changed files")
+	patchPath := flag.String("patch", "", "Path to patch of changed files")
 	flag.Parse()
-	if flag.NArg() != 0 {
-		log.Fatalf("Unexpected argument.")
-	}
-	// Read Tricium input FILES data.
-	input := &tricium.Data_Files{}
-	if err := tricium.ReadDataType(*inputDir, input); err != nil {
-		log.Fatalf("Failed to read FILES data: %v", err)
-	}
-	log.Printf("Read FILES data.")
-
-	results := &tricium.Data_Results{}
-
-	// Only add .xml files to filePaths.
+	// This is a temporary way for us to get recipes to work without breaking the current analyzer.
 	var filePaths []string
-	for _, file := range input.Files {
-		if !file.IsBinary && filepath.Ext(file.Path) == ".xml" {
-			filePaths = append(filePaths, filepath.Join(*inputDir, file.Path))
+	if *prevDir != "" {
+		filePaths = flag.Args()
+	} else {
+		// Read Tricium input FILES data.
+		input := &tricium.Data_Files{}
+		if err := tricium.ReadDataType(*inputDir, input); err != nil {
+			log.Fatalf("Failed to read FILES data: %v", err)
 		}
+		log.Printf("Read FILES data.")
+		// Only add .xml files to filePaths.
+		for _, file := range input.Files {
+			if !file.IsBinary && filepath.Ext(file.Path) == ".xml" {
+				filePaths = append(filePaths, file.Path)
+			}
+		}
+		// Return early if no .xml files were modified.
+		if len(filePaths) == 0 {
+			return
+		}
+		// Set up the temporary directory where we'll put original files.
+		// The temporary directory should be cleaned up before exiting.
+		tempDir, err := ioutil.TempDir(*inputDir, "get-original-file")
+		if err != nil {
+			log.Fatalf("Failed to setup temporary directory: %v", err)
+		}
+		defer func() {
+			if err = os.RemoveAll(tempDir); err != nil {
+				log.Fatalf("Failed to clean up temporary directory %q: %v", tempDir, err)
+			}
+		}()
+		log.Printf("Created temporary directory %q.", tempDir)
+		*patchPath = input.Patch
+		*prevDir = filepath.Join(tempDir, *inputDir)
+		// Original files will be put into prevDir.
+		getOriginalFiles(filePaths, *inputDir, *prevDir, *patchPath)
 	}
-
-	// Return early if no .xml files were modified.
-	if len(filePaths) == 0 {
-		return
-	}
-
-	filesChanged, err := getDiffsPerFile(filepath.Join(*inputDir, input.Patch))
+	filesChanged, err := getDiffsPerFile(filepath.Join(*inputDir, *patchPath))
 	if err != nil {
 		log.Fatalf("Failed to get diffs per file: %v", err)
 	}
 
-	// Set up the temporary directory where we'll put original files.
-	// The temporary directory should be cleaned up before exiting.
-	tempDir, err := ioutil.TempDir(*inputDir, "get-original-file")
-	if err != nil {
-		log.Fatalf("Failed to setup temporary directory: %v", err)
-	}
-	defer func() {
-		if err = os.RemoveAll(tempDir); err != nil {
-			log.Fatalf("Failed to clean up temporary directory %q: %v", tempDir, err)
-		}
-	}()
-	log.Printf("Created temporary directory %q.", tempDir)
-
-	// Original files will be put into tempDir.
-	getOriginalFiles(filePaths, tempDir, filepath.Join(*inputDir, input.Patch))
-
+	results := &tricium.Data_Results{}
 	for _, filePath := range filePaths {
-		results.Comments = append(results.Comments, analyzeFile(filePath, tempDir, filesChanged)...)
+		results.Comments = append(results.Comments, analyzeFile(filePath, *inputDir, *prevDir, filesChanged)...)
 	}
 
 	// Write Tricium RESULTS data.
@@ -234,18 +234,18 @@ func getDiffsPerFile(patchPath string) (*diffsPerFile, error) {
 }
 
 // getOriginalFiles gets files in parent commit, before the patch, and puts them in tempDir.
-func getOriginalFiles(filePaths []string, tempDir string, patchPath string) {
+func getOriginalFiles(filePaths []string, inputDir string, prevDir string, patchPath string) {
 	filesToCopy := append(filePaths, patchPath)
 	for _, filePath := range filesToCopy {
-		tempPath := filepath.Join(tempDir, filePath)
+		tempPath := filepath.Join(prevDir, filePath)
 		// Note: Must use filepath.Dir rather than path.Dir to be compatible with Windows.
 		if err := os.MkdirAll(filepath.Dir(tempPath), os.ModePerm); err != nil {
 			log.Fatalf("Failed to create dirs for file: %v", err)
 		}
-		copyFile(filePath, tempPath)
+		copyFile(filepath.Join(inputDir, filePath), tempPath)
 	}
 	// Only apply patch if patch is not empty.
-	fi, err := os.Stat(patchPath)
+	fi, err := os.Stat(filepath.Join(inputDir, patchPath))
 	if err != nil {
 		log.Fatalf("Failed to get file info for patch %s: %v", patchPath, err)
 	}
@@ -254,7 +254,7 @@ func getOriginalFiles(filePaths []string, tempDir string, patchPath string) {
 		cmds = append(cmds, exec.Command("git", "apply", "-p1", "--reverse", patchPath))
 		for _, c := range cmds {
 			var stderr bytes.Buffer
-			c.Dir = tempDir
+			c.Dir = prevDir
 			c.Stderr = &stderr
 			log.Printf("Running cmd: %s", c.Args)
 			if err := c.Run(); err != nil {
@@ -274,19 +274,20 @@ func copyFile(sourceFile string, destFile string) {
 	}
 }
 
-func analyzeFile(inputPath string, tempDir string, filesChanged *diffsPerFile) []*tricium.Data_Comment {
-	log.Printf("ANALYZING File: %s", inputPath)
+func analyzeFile(filePath string, inputDir string, prevDir string, filesChanged *diffsPerFile) []*tricium.Data_Comment {
+	log.Printf("ANALYZING File: %s", filePath)
 	var allComments []*tricium.Data_Comment
+	inputPath := filepath.Join(inputDir, filePath)
 	f := openFileOrDie(inputPath)
 	defer closeFileOrDie(f)
 	// Analyze added lines in file (if any)
-	comments, addedHistograms, newNamespaces, namespaceLineNums := analyzeChangedLines(bufio.NewScanner(f), inputPath, filesChanged.addedLines[inputPath], ADDED)
+	comments, addedHistograms, newNamespaces, namespaceLineNums := analyzeChangedLines(bufio.NewScanner(f), inputPath, filesChanged.addedLines[filePath], ADDED)
 	allComments = append(allComments, comments...)
 	// Analyze removed lines in file (if any)
-	tempPath := filepath.Join(tempDir, inputPath)
+	tempPath := filepath.Join(prevDir, filePath)
 	oldFile := openFileOrDie(tempPath)
 	defer closeFileOrDie(oldFile)
-	_, removedHistograms, oldNamespaces, _ := analyzeChangedLines(bufio.NewScanner(oldFile), tempPath, filesChanged.removedLines[inputPath], REMOVED)
+	_, removedHistograms, oldNamespaces, _ := analyzeChangedLines(bufio.NewScanner(oldFile), tempPath, filesChanged.removedLines[filePath], REMOVED)
 	// Identify any removed histograms
 	allComments = append(allComments, findRemovedHistograms(inputPath, addedHistograms, removedHistograms)...)
 	allComments = append(allComments, findAddedNamespaces(inputPath, newNamespaces, oldNamespaces, namespaceLineNums)...)
